changed httpd.include.conf file name to scripts/apache-httpd.include.conf

+++ scripts/suse-otrs.spec
@@ -2,7 +2,7 @@
 # RPM spec file for SuSE Linux of the OTRS package
 # Copyright (C) 2002-2003 Martin Edenhofer <bugs+rpm@otrs.org>
 # --
-# $Id: suse-otrs.spec,v 1.27 2003-01-05 23:51:56 martin Exp $
+# $Id: suse-otrs.spec,v 1.28 2003-01-09 20:55:37 martin Exp $
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see 
 # the enclosed file COPYING for license information (GPL). If you 
@@ -18,7 +18,7 @@ Version:      0.5
 Copyright:    GNU GENERAL PUBLIC LICENSE Version 2, June 1991
 Group:        Applications/Mail
 Provides:     otrs 
-Requires:     perl perl-DBI perl-GD perl-Net-DNS perl-Digest-MD5 apache mod_perl mysql mysql-client perl-Msql-Mysql-modules mysql-shared fetchmail procmail
+Requires:     perl perl-DBI perl-GD perl-Net-DNS perl-Digest-MD5 apache mod_perl mysql mysql-client perl-Msql-Mysql-modules mysql-shared fetchmail procmail perl-IO-stringy
 Autoreqprov:  on
 Release:      BETA7
 Source0:      otrs-%{version}-%{release}.tar.bz2
@@ -111,7 +111,7 @@ SuSE series: ap
 cp Kernel/Config.pm.dist Kernel/Config.pm
 cd Kernel/Config/ && for foo in *.dist; do cp $foo `basename $foo .dist`; done && cd ../../
 # copy all crontab dist files
-for foo in var/cron/*.dist; do cp $foo var/cron/`basename $foo .dist`; done
+for foo in var/cron/*.dist; do mv $foo var/cron/`basename $foo .dist`; done
 
 %install
 # delete old RPM_BUILD_ROOT
@@ -157,13 +157,20 @@ chown wwwrun /opt/otrs/Kernel/Config.pm
 # sysconfig
 %{fillup_and_insserv -s otrs START_OTRS}
 
-# add suse-httpd.include.conf to apache.rc.config
-APACHERC=/etc/sysconfig/apache
-
-OTRSINCLUDE=/opt/otrs/scripts/suse-httpd.include.conf
-sed 's+^HTTPD_CONF_INCLUDE_FILES=.*$+HTTPD_CONF_INCLUDE_FILES='$OTRSINCLUDE'+' \
-$APACHERC > /tmp/apache.rc.config.tmp && mv /tmp/apache.rc.config.tmp $APACHERC 
-
+# add httpd.include.conf to /etc/sysconfig/apache
+if test -e /etc/sysconfig/apache; then
+    OTRSINCLUDE=/opt/otrs/scripts/apache-httpd.include.conf
+    APACHERC=/etc/sysconfig/apache
+    sed 's+^HTTPD_CONF_INCLUDE_FILES=.*$+HTTPD_CONF_INCLUDE_FILES='$OTRSINCLUDE'+' \
+    $APACHERC > /tmp/apache.rc.config.tmp && mv /tmp/apache.rc.config.tmp $APACHERC 
+fi
+# add suse-httpd.include.conf to /etc/sysconfig/apache2
+if test -e /etc/sysconfig/apache2; then
+    OTRSINCLUDE=/opt/otrs/scripts/apache2-httpd.include.conf
+    APACHERC=/etc/sysconfig/apache2
+    sed 's+^APACHE_CONF_INCLUDE_FILES=.*$+APACHE_CONF_INCLUDE_FILES='$OTRSINCLUDE'+' \
+    $APACHERC > /tmp/apache.rc.config.tmp && mv /tmp/apache.rc.config.tmp $APACHERC 
+fi
 # note
 echo ""
 echo "Next steps: "
